refactor: refactored controller and it's test code
