+++ src/enum/httpStatus.ts
@@ -12,3 +12,8 @@ export enum HttpStatusMessage {
   NOT_FOUND = 'NOT_FOUND',
   INTERNAL_SERVER = 'INTERNAL_SERVER_ERROR',
 };
+
+export enum ResponseMessage {
+  MISSING_FILE = "Csv file is missing",
+  WRONG_METHOD_TYPE = 'Wrong method type, expected method type is POST'
+}
