+++ src/router/index.ts
@@ -1,6 +1,6 @@
 import { IncomingMessage, ServerResponse } from 'http';
 import { RESPONSE_HEADER, ROUTES } from '../constants';
-import { organOrderController } from '../controllers/organOrder';
+import { organOrderController } from '../controllers/organOrderController';
 import { sendResponse } from '../utils/sendResponse';
 import { HttpStatusCode, HttpStatusMessage } from '../enum/httpStatus';
